+++ integration/env.go
@@ -6,11 +6,11 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"os/exec"
 	"path"
 	"path/filepath"
 	"strconv"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
@@ -19,9 +19,9 @@ import (
 
 	"github.com/ignite/cli/ignite/chainconfig"
 	"github.com/ignite/cli/ignite/pkg/cosmosfaucet"
+	"github.com/ignite/cli/ignite/pkg/gocmd"
 	"github.com/ignite/cli/ignite/pkg/gomodulepath"
 	"github.com/ignite/cli/ignite/pkg/httpstatuschecker"
-	"github.com/ignite/cli/ignite/pkg/xexec"
 	"github.com/ignite/cli/ignite/pkg/xurl"
 )
 
@@ -30,28 +30,15 @@ const (
 )
 
 var (
-	isCI, _   = strconv.ParseBool(os.Getenv("CI"))
-	IgniteApp string
-)
+	// IgniteApp hold the location of the ignite binary used in the integration
+	// tests. The binary is compiled the first time the env.New() function is
+	// invoked.
+	IgniteApp = path.Join(os.TempDir(), "ignite-tests", "ignite")
 
-func init() {
-	wd, _ := os.Getwd()
-	_, appPath, err := gomodulepath.Find(wd)
-	if err != nil {
-		panic(err)
-	}
-	// Build the ignite binary
-	tmp, err := os.MkdirTemp("", "integration-bin")
-	if err != nil {
-		panic(err)
-	}
-	IgniteApp = path.Join(tmp, "ignite")
-	command := exec.Command("go", "build",
-		"-o", IgniteApp, appPath+"/ignite/cmd/ignite")
-	if err := command.Run(); err != nil {
-		panic(err)
-	}
-}
+	isCI, _           = strconv.ParseBool(os.Getenv("CI"))
+	mainPackage       = path.Join("ignite", "cmd", "ignite")
+	compileBinaryOnce sync.Once
+)
 
 // Env provides an isolated testing environment and what's needed to
 // make it possible.
@@ -68,12 +55,29 @@ func New(t *testing.T) Env {
 		ctx: ctx,
 	}
 	t.Cleanup(cancel)
-
-	if !xexec.IsCommandAvailable(IgniteApp) {
-		t.Fatal("ignite needs to be installed")
-	}
-
+	compileBinaryOnce.Do(func() {
+		compileBinary(ctx)
+	})
 	return e
+}
+
+func compileBinary(ctx context.Context) {
+	wd, err := os.Getwd()
+	if err != nil {
+		panic(fmt.Sprintf("unable to get working dir: %v", err))
+	}
+	_, appPath, err := gomodulepath.Find(wd)
+	if err != nil {
+		panic(fmt.Sprintf("unable to read go module path: %v", err))
+	}
+	var (
+		output, binary = filepath.Split(IgniteApp)
+		path           = path.Join(appPath, mainPackage)
+	)
+	err = gocmd.BuildPath(ctx, output, binary, path, nil)
+	if err != nil {
+		panic(fmt.Sprintf("error while building binary: %v", err))
+	}
 }
 
 // SetCleanup registers a function to be called when the test (or subtest) and all its